changed based on feedback, added strict version
