+++ helix-core/src/main/java/org/apache/helix/tools/ClusterVerifiers/ZkHelixClusterVerifier.java
@@ -198,10 +198,9 @@ public abstract class ZkHelixClusterVerifier
         }
         TimeUnit.MILLISECONDS.sleep(period);
       } while ((System.currentTimeMillis() - start) <= timeout);
-      LOG.error("verifier timeout out with timeout {},. stack trace {} ",
-          timeout, Arrays.asList(Thread.currentThread().getStackTrace()));
+      LOG.error("verifier timeout out with timeout {}", timeout);
     } catch (Exception e) {
-      LOG.error("Exception in verifier {}", e);
+      LOG.error("Exception in verifier", e);
     }
     return false;
   }
@@ -238,13 +237,12 @@ public abstract class ZkHelixClusterVerifier
           // make a final try if timeout
           success = verifyState();
           if (!success) {
-            LOG.error("verifyByCallback failed due to timeout, with stack trace {}",
-                Arrays.asList(Thread.currentThread().getStackTrace()));
+            LOG.error("verifyByCallback failed due to timeout");
           }
         }
       }
     } catch (Exception e) {
-      LOG.error("Exception in verifier {}", e);
+      LOG.error("Exception in verifier", e);
     }
 
     // clean up
